[EGD-7134] Network setting cleanup and fixes
Network setting cleanup. Added missing sim cards window.
Fixed option texts. Disabled not working options in UI.
