+++ module-apps/application-settings/models/network/NewApnModel.hpp
@@ -3,12 +3,10 @@
 
 #pragma once
 
-#include "application-settings/data/SettingsItemData.hpp"
-#include "application-settings/widgets/ApnListItem.hpp"
-#include "application-settings/widgets/SettingsStyle.hpp"
-#include "InternalModel.hpp"
-#include "Application.hpp"
+#include <application-settings/widgets/network/ApnListItem.hpp>
 
+#include <InternalModel.hpp>
+#include <Application.hpp>
 #include <ListItemProvider.hpp>
 
 class NewApnModel : public app::InternalModel<gui::ApnListItem *>, public gui::ListItemProvider
