+++ module-apps/application-settings/widgets/network/ApnListItem.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "application-settings/data/SettingsItemData.hpp"
+#include <application-settings/data/SettingsItemData.hpp>
 
 namespace gui
 {
