+++ module-apps/application-settings/windows/network/ApnOptionsWindow.hpp
@@ -3,8 +3,8 @@
 
 #pragma once
 
-#include "application-settings/models/ApnSettingsModel.hpp"
-#include "BaseSettingsWindow.hpp"
+#include <application-settings/models/network/ApnSettingsModel.hpp>
+#include <application-settings/windows/BaseSettingsWindow.hpp>
 
 #include <service-cellular/PacketDataTypes.hpp>
