+++ module-apps/apps-common/options/OptionStyle.hpp
@@ -40,8 +40,6 @@ namespace gui::option
         inline constexpr gui::Length option_left_margin       = 10;
         inline constexpr gui::Length option_right_margin      = 10;
         inline constexpr gui::Length option_bottom_margin     = style::margins::big;
-        inline constexpr gui::Length option_right_min_size    = 150;
-        inline constexpr gui::Length option_rightbig_min_size = 350;
         inline constexpr uint32_t optionsListTopMargin        = 10U;
         inline constexpr uint32_t optionsListX                = style::window::default_left_margin;
         inline constexpr uint32_t optionsListY = style::window::default_vertical_pos + optionsListTopMargin;
