+++ module-apps/apps-common/options/type/OptionSetting.cpp
@@ -65,14 +65,13 @@ namespace gui::option
         case SettingRightItem::Text: {
             auto optionTextRight = new TextFixedSize(optionBodyHBox, 0, 0, 0, 0);
             optionTextRight->setUnderline(false);
-            optionTextRight->setMinimumSize((textOnRightIsSmall) ? gui::option::window::option_right_min_size
-                                                                 : gui::option::window::option_rightbig_min_size,
-                                            style::window::label::big_h);
+            optionTextRight->setFont((textOnRightIsSmall) ? style::window::font::verysmall
+                                                          : style::window::font::medium);
+            optionTextRight->setMinimumWidthToFitText(textOnRight);
+            optionTextRight->setMinimumHeight(style::window::label::big_h);
             optionTextRight->setAlignment(
                 gui::Alignment(gui::Alignment::Horizontal::Right, gui::Alignment::Vertical::Center));
             optionTextRight->setMargins(Margins(0, 0, window::option_right_margin, 0));
-            optionTextRight->setFont((textOnRightIsSmall) ? style::window::font::medium
-                                                          : style::window::font::verysmall);
             optionTextRight->setRichText(textOnRight);
             break;
         }
